chore(TS): migrate parser (#8122)
