+++ src/parser/getSvgRegex.ts
@@ -1,0 +1,5 @@
+//@ts-nocheck
+
+export function getSvgRegex(arr) {
+  return new RegExp('^(' + arr.join('|') + ')\\b', 'i');
+}
