+++ src/parser/skewMatrix.ts
@@ -1,0 +1,7 @@
+//@ts-nocheck
+import { degreesToRadians } from '../util';
+
+
+export function skewMatrix(matrix, args, pos) {
+    matrix[pos] = Math.tan(degreesToRadians(args[0]));
+}
