+++ src/util/path.ts
@@ -1,4 +1,5 @@
 //@ts-nocheck
+import { commaWsp, rePathCommand } from "../parser/constants";
 import { Point } from '../point.class';
 
 (function(global) {
@@ -682,10 +683,10 @@ import { Point } from '../point.class';
         coords = [],
         currentPath,
         parsed,
-        re = fabric.rePathCommand,
+        re = rePathCommand,
         rNumber = '[-+]?(?:\\d*\\.\\d+|\\d+\\.?)(?:[eE][-+]?\\d+)?\\s*',
-        rNumberCommaWsp = '(' + rNumber + ')' + fabric.commaWsp,
-        rFlagCommaWsp = '([01])' + fabric.commaWsp + '?',
+        rNumberCommaWsp = '(' + rNumber + ')' + commaWsp,
+        rFlagCommaWsp = '([01])' + commaWsp + '?',
         rArcSeq = rNumberCommaWsp + '?' + rNumberCommaWsp + '?' + rNumberCommaWsp + rFlagCommaWsp + rFlagCommaWsp +
           rNumberCommaWsp + '?(' + rNumber + ')',
         regArcArgumentSequence = new RegExp(rArcSeq, 'g'),
